Refactor: CommunicationDisplay에서 Conversation으로
