+++ src/utils/Chat.ts
@@ -1,4 +1,4 @@
-import { MutableRefObject } from "react";
+import { MutableRefObject, useEffect, useRef } from "react";
 
 export const smoothScrollToBottom = (
 	element: MutableRefObject<HTMLDivElement | null>
@@ -10,4 +10,14 @@ export const smoothScrollToBottom = (
 		behavior: "smooth",
 		block: "end",
 	});
+};
+
+export const useSmoothScrollToBottom = () => {
+	const divRef = useRef<HTMLDivElement | null>(null);
+
+	useEffect(() => {
+		smoothScrollToBottom(divRef);
+	}, [divRef]);
+
+	return divRef;
 };
